requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
plotly==5.14.1
dash==2.9.3
networkx==3.1
google-genai